\name{TADreg-package}
\alias{TADreg-package}
\alias{TADreg}
\docType{package}
\title{
\packageTitle{TADreg}
}
\description{
\packageDescription{TADreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TADreg}
\packageIndices{TADreg}
To use TADreg, see the R Markdown html file which shows examples to run the different functions (SIM, DIM and PIM) from the package.
}
\author{
\packageAuthor{TADreg}

Maintainer: \packageMaintainer{TADreg}
}
%\references{
%~~ Literature or other references for background information ~~
%}
